Move ValueMap to mod file
